Changed datatype from sbyte to short in anime-lists XML parser
This prevents overflow errors with some new animes having a offset greater than -127

+++ MediaBrowser.Plugins.Anime/anime-lists-net/AnimelistAnime.cs
@@ -44,7 +44,7 @@ namespace AnimeLists
 
         /// <remarks />
         [XmlAttribute("episodeoffset")]
-        public byte EpisodeOffset { get; set; }
+        public short EpisodeOffset { get; set; }
 
         /// <remarks />
         [XmlIgnore]
